feat(datagrid-text-filter-web): removing logic from rendering component
